fix: 당첨금액 타입 long 으로 변경

+++ src/main/java/lotto/domain/LottoCalculator.java
@@ -2,18 +2,18 @@ package lotto.domain;
 
 public class LottoCalculator {
 
-	public static int getLottoCount(int money) {
+	public static int getLottoCount(long money) {
 		validMoney(money);
-		return money / Lotto.LOTTO_PRICE;
+		return (int) (money / Lotto.LOTTO_PRICE);
 	}
 
-	private static void validMoney(int money) {
+	private static void validMoney(long money) {
 		if (money < 1000) {
 			throw new IllegalArgumentException("로또을 사려면 1000원 이상 있어야 합니다.");
 		}
 	}
 
-	public static double getRateOfReturn(int initMoney, int getMoney) {
+	public static double getRateOfReturn(long initMoney, long getMoney) {
 		double rate = (double) getMoney / initMoney;
 		return (Math.floor(rate * 100) / 100.0);
 	}
